Restructure the handler/helper classes code for build/test
